2025.5.26

+++ include/SignalPool.h
@@ -3,10 +3,10 @@
 
 class SignalPool {
 private:
-    std :: map<uint64_t, std :: map<std :: string, int> > pool;
+    std :: map<uint64_t, std :: map<std :: string, uint64_t> > pool;
 public:
-    void add(uint64_t receiver, const std :: string &signal, int content = 0);
+    void add(uint64_t receiver, const std :: string &signal, uint64_t content = 0);
     bool contains(uint64_t owner, const std :: string &signal) const;
     void del(uint64_t owner, const std :: string &signal);
-    int query(uint64_t owner, const std :: string &signal) const;
+    uint64_t query(uint64_t owner, const std :: string &signal) const;
 };
