+++ src/SignalPool.cpp
@@ -1,6 +1,6 @@
 #include "SignalPool.h"
 
-void SignalPool :: add(uint64_t receiver, const std :: string &signal, int content) {
+void SignalPool :: add(uint64_t receiver, const std :: string &signal, uint64_t content) {
     pool[receiver][signal] = content;
 }
 bool SignalPool :: contains(uint64_t owner, const std :: string &signal) const {
@@ -11,7 +11,7 @@ void SignalPool :: del(uint64_t owner, const std :: string &signal) {
         pool[owner].erase(signal);
     }
 }
-int SignalPool :: query(uint64_t owner, const std :: string &signal) const {
+uint64_t SignalPool :: query(uint64_t owner, const std :: string &signal) const {
     if(!contains(owner, signal)) return -1;
     return pool.at(owner).at(signal);
 }
